# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a list of courses",
                "description": "Display a listing of the resource.",
                "operationId": "02eeda67042dba2d22a517f26fb8eaf1",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter courses by year",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Filter courses by name or teacher name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "semester",
                        "in": "query",
                        "description": "Filter courses by semester",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of courses per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CourseResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/courses/{course}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a specific course by ID",
                "description": "Display the specified resource.",
                "operationId": "c296c0c8738577d66bdd1332e8684906",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "slug of the course to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/courses/{course}/teachers": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get teachers for a specific course",
                "description": "Get the teachers associated with a specific course.",
                "operationId": "b19e8a6983d2e47dda47c4f32180065e",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "slug of the course to retrieve teachers for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TeacherResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/courses/{course}/sessions": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get sessions for a specific course",
                "description": "Get the sessions associated with a specific course.",
                "operationId": "0babb65b20579b2b3e7bb005aefcaf78",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "slug of the course to retrieve sessions for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userSaved",
                        "in": "query",
                        "description": "Filter sessions saved by the authenticated user",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CourseSessionResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/courses/{course}/save": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Add a course to the user's saved list",
                "description": "Add a course to the authenticated user's saved list.",
                "operationId": "48e38feff707d58c0dcc105d4b77658c",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "The ID of the course to save",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course successfully added to saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{course}/unsave": {
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Remove a course from the user's saved list",
                "description": "Remove a course from the authenticated user's saved list.",
                "operationId": "f4a15ebbde6b44be6fb97a5ff720cfb4",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "The ID of the course to unsave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course successfully removed from saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/grouped-by-year": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get courses grouped by year",
                "description": "Get courses grouped by year.",
                "operationId": "49f948f69225650ae0ca3d6d00693dc4",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CourseResource": {
                "title": "Course Resource",
                "description": "Course resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the course",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the course",
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "summary": {
                        "description": "A brief summary of the course",
                        "type": "string",
                        "example": "Learn the basics of Laravel framework"
                    },
                    "studentsCount": {
                        "description": "The number of students enrolled in the course",
                        "type": "integer",
                        "example": 25
                    },
                    "createdAt": {
                        "description": "The date and time when the course was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "teachers": {
                        "description": "List of teachers associated with the course",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeacherResource"
                        }
                    },
                    "primaryImage": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "images": {
                        "description": "List of images associated with the course",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    },
                    "quizezPercent": {
                        "description": "The percentage of quizzes completed by the student",
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "lecturesPercent": {
                        "description": "The percentage of lectures completed by the student",
                        "type": "number",
                        "format": "float",
                        "example": 60
                    },
                    "type": {
                        "description": "The type of the course",
                        "type": "string",
                        "example": "online"
                    },
                    "discount": {
                        "description": "The discount applied to the course",
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "price": {
                        "description": "The price of the course",
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "rating": {
                        "description": "The rating of the course",
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "teachersCount": {
                        "description": "The number of teachers associated with the course",
                        "type": "integer",
                        "example": 3
                    },
                    "year": {
                        "description": "The year the course was created",
                        "type": "integer",
                        "example": 2023
                    },
                    "time": {
                        "description": "The duration of the course in hours",
                        "type": "integer",
                        "example": 10
                    },
                    "quizezCount": {
                        "description": "The number of quizzes in the course",
                        "type": "integer",
                        "example": 5
                    },
                    "filesCount": {
                        "description": "The number of files associated with the course",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "CourseSessionResource": {
                "title": "Course Session Resource",
                "description": "Course Session resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the course session",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the course session",
                        "type": "string"
                    },
                    "note": {
                        "description": "Additional notes for the course session",
                        "type": "string"
                    },
                    "time": {
                        "description": "The time of the course session",
                        "type": "string"
                    },
                    "like": {
                        "description": "The number of likes for the course session",
                        "type": "integer"
                    },
                    "dislike": {
                        "description": "The number of dislikes for the course session",
                        "type": "integer"
                    },
                    "type": {
                        "description": "The type of the course session",
                        "type": "string"
                    },
                    "isFree": {
                        "description": "Indicates if the course session is free",
                        "type": "boolean"
                    },
                    "teacher": {
                        "$ref": "#/components/schemas/TeacherResource"
                    },
                    "course": {
                        "$ref": "#/components/schemas/CourseResource"
                    },
                    "video": {
                        "description": "The video media associated with the course session",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    },
                    "image": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "currentUserProgress": {
                        "$ref": "#/components/schemas/UserProgressResource"
                    }
                },
                "type": "object"
            },
...
            "TeacherResource": {
                "title": "Teacher Resource",
                "description": "Teacher resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the teacher",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the teacher",
                        "type": "string"
                    },
                    "rate": {
                        "description": "The rating of the teacher",
                        "type": "number",
                        "format": "float"
                    },
                    "image": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "createdAt": {
                        "description": "The date when the teacher was created",
                        "type": "string",
                        "format": "date"
                    },
                    "reviewsCount": {
                        "description": "The number of reviews for the teacher",
                        "type": "integer"
                    },
                    "studentsCount": {
                        "description": "The number of students for the teacher",
                        "type": "integer"
                    },
                    "coursesCount": {
                        "description": "The number of courses for the teacher",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UserProgressResource": {
                "title": "User Progress Resource",
                "description": "User progress resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the user progress record",
                        "type": "integer"
                    },
                    "complete": {
                        "description": "Indicates if the user has completed the progress",
                        "type": "boolean"
                    },
                    "lastTime": {
                        "description": "The last time the user updated the progress",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}